Main UI initial functionality
