+++ src/main/java/Main/Main.java
@@ -1,14 +1,12 @@
 package Main;
 
-import FlashcardProgram.FlashcardSystem;
-
 public class Main {
     public static void main(String[] args) {
-        FlashcardSystem fs = new FlashcardSystem();
-        fs.displayMainMenu();
+        // FlashcardSystem fs = new FlashcardSystem();
+        // fs.displayMainMenu();
         // TODO: properly use modules instead of this hack for JavaFX
         // LearningSessionUI.main(args);
-        // UI.MainUI.main(args);
+        UI.Main.main(args);
 
     }
 }
