--- src/main/java/UI/MainUIController.java
@@ -1,121 +1,0 @@
-package UI;
-
-import javafx.application.Platform;
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.Node;
-import javafx.scene.control.MenuButton;
-import javafx.scene.control.TextField;
-import javafx.stage.Modality;
-import javafx.stage.Stage;
-
-import java.io.IOException;
-
-public class MainUIController {
-    @FXML private MenuButton deckSelect;
-    @FXML private MenuButton sessionSelect;
-    @FXML private TextField deckName;
-
-    /**
-     * Opens Deck creation menu
-     * @param e action event on click
-     * @throws IOException if create-deck-view.fxml is not found
-     */
-    @FXML
-    protected void onCreateDeckButtonClick(ActionEvent e) throws IOException {
-        Parent createDeckParent = FXMLLoader.load(getClass().getResource("/create-deck-view.fxml"));
-        Scene createDeckScene = new Scene(createDeckParent);
-
-        Stage stage = (Stage)((Node)e.getSource()).getScene().getWindow();
-        Stage popUp = new Stage();
-        popUp.setTitle("Create New Deck");
-        popUp.setScene((createDeckScene));
-
-        popUp.initModality(Modality.WINDOW_MODAL);
-        popUp.initOwner(stage);
-        popUp.setX(stage.getX()+20);
-        popUp.setY(stage.getY()+20);
-
-        popUp.show();
-    }
-
-    @FXML
-    protected void onCreateDeckSubmit(ActionEvent e) {
-        //TODO: create deck with deckName.getText();
-        onBackButtonClick(e);
-    }
-
-    /**
-     * Opens Deck study menu
-     * @param e action event on click
-     * @throws IOException if study-deck-view.fxml is not found
-     */
-    @FXML
-    protected void onStudyDeckButtonClick(ActionEvent e) throws IOException {
-        Parent studyDeckParent = FXMLLoader.load(getClass().getResource("/study-deck-view.fxml"));
-        Scene studyDeckScene = new Scene(studyDeckParent);
-
-        Stage stage = (Stage)((Node)e.getSource()).getScene().getWindow();
-        Stage popUp = new Stage();
-        popUp.setTitle("Initialize New Session");
-        popUp.setScene((studyDeckScene));
-
-        popUp.initModality(Modality.WINDOW_MODAL);
-        popUp.initOwner(stage);
-        popUp.setX(stage.getX()+20);
-        popUp.setY(stage.getY()+20);
-
-        popUp.show();
-    }
-
-    /**
-     * Opens study session application
-     */
-    @FXML
-    protected void onStartSessionSubmit() {
-        Platform.runLater(
-                () -> {
-                    try {
-                        new LearningSessionUI().start(new Stage());
-                    } catch (Exception e) {
-                        e.printStackTrace();
-                    }
-                }
-        );
-    }
-
-    /**
-     * Opens Deck edit menu
-     * @param e action event on click
-     * @throws IOException if edit-deck-view.fxml is not found
-     */
-    @FXML
-    protected void onEditDeckButtonClick(ActionEvent e) throws IOException {
-        Parent editDeckParent = FXMLLoader.load(getClass().getResource("/edit-deck-view.fxml"));
-        Scene editDeckScene = new Scene(editDeckParent);
-        Stage stage = (Stage)((Node)e.getSource()).getScene().getWindow();
-        stage.setScene(editDeckScene);
-        stage.show();
-    }
-
-    /**
-     * Returns to previous stage
-     * @param e action event on click
-     */
-    @FXML
-    protected void onBackButtonClick(ActionEvent e) {
-        Stage stage = (Stage)((Node)e.getSource()).getScene().getWindow();
-        stage.close();
-    }
-
-    /**
-     * Closes the program
-     */
-    @FXML
-    protected void onLogoutButtonClick() {
-        Platform.exit();
-    }
-}
