+++ src/main/resources/create-deck-view.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.control.TextField?>
 
 <BorderPane prefHeight="400.0" prefWidth="600.0" xmlns:fx="http://javafx.com/fxml"
-            fx:controller="UI.MainUIController">
+            fx:controller="UI.MainUI">
     <top>
         <Label BorderPane.alignment="TOP_CENTER" text="Flashcards!"/>
     </top>
