+++ src/main/resources/study-deck-view.fxml
@@ -6,9 +6,11 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.layout.BorderPane?>
 
+<?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.MenuButton?>
+
 <BorderPane prefHeight="400.0" prefWidth="600.0" xmlns:fx="http://javafx.com/fxml"
-            fx:controller="UI.MainUIController">
+            fx:controller="UI.MainUI">
     <top>
         <Label BorderPane.alignment="TOP_CENTER" text="Flashcards!"/>
     </top>
@@ -17,7 +19,7 @@
             <padding>
                 <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
             </padding>
-            <MenuButton fx:id="deckSelect" text="Select Deck"/>
+            <ComboBox fx:id="deckSelect"/>
             <MenuButton fx:id="sessionSelect" text="Select Session"/>
             <Button onAction="#onStartSessionSubmit" fx:id="studyButton" text="Start Session"/>
         </VBox>
